requirements.txt
biopython
django-registration
south
django==1.4
selenium
MySQL-python
simplejson
numpy
cairocffi